# backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DiscoveryShop Marketplace API",
    "version": "1.0.0",
    "description": "Professional API documentation for DiscoveryShop - Full-stack e-commerce marketplace platform with AI-powered discovery, dual-role accounts, real-time chat, and comprehensive admin controls.",
    "contact": {
      "name": "API Support",
      "email": "support@discoveryshop.local"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development Server"
    },
    {
      "url": "https://api.discoveryshop.com",
      "description": "Production Server"
    }
  ],
  "paths": {
...
    "/api/v1/admin/products/{product_id}/approve": {
      "post": {
        "tags": ["Admin"],
        "summary": "Approve product",
        "description": "Approve pending product for listing. Requires admin role.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {"type": "integer"}
          }
        ],
        "responses": {
          "200": {
            "description": "Product approved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/admin/products/{product_id}/reject": {
      "post": {
        "tags": ["Admin"],
        "summary": "Reject product",
        "description": "Reject pending product with reason. Requires admin role.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {"type": "integer"}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {"type": "string"}
                },
                "required": ["reason"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product rejected successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from login endpoint"
      }
    },
...
  },
...
}